feat: retrieve info about chunks from `streamer_message`
