+++ src/rpc.rs
@@ -7,12 +7,10 @@ use near_jsonrpc_primitives::types::query::QueryResponseKind;
 use near_primitives::hash::CryptoHash;
 use near_primitives::transaction::{Action, FunctionCallAction, Transaction, TransactionV0};
 use near_primitives::types::{BlockHeight, BlockReference};
-use near_primitives::views::{ActionView, ChunkHeaderView, FinalExecutionStatus};
+use near_primitives::views::{ChunkHeaderView, FinalExecutionStatus};
 use near_sdk::borsh::{self, BorshDeserialize, BorshSerialize};
-use near_sdk::json_types::U128;
 use near_sdk::AccountId;
 use std::collections::HashSet;
-use std::str::FromStr;
 use std::time::Duration;
 
 use self::error::CommitTx;
@@ -43,7 +41,7 @@ const RETRIES_COUNT: u8 = 10;
 
 /// Transaction action methods allowed for output parsing and
 /// get `predecessor_account_id`
-const ACTION_METHODS: &[&str] = &[
+pub const ACTION_METHODS: &[&str] = &[
     "ft_transfer",
     "deposit",
     "ft_transfer_call",
@@ -146,253 +144,6 @@ impl Client {
             block.header.hash,
             block.header.prev_hash,
         ))
-    }
-
-    /// Get action output for chunk transaction (including receipt output)
-    /// It includes: Accounts
-    pub fn get_actions_data(&mut self, actions: Vec<ActionView>) -> ActionResult {
-        let mut result = ActionResult::default();
-
-        for action in actions {
-            // Check action method and filter it
-            if let ActionView::FunctionCall {
-                method_name, args, ..
-            } = action
-            {
-                if ACTION_METHODS.contains(&method_name.as_str()) {
-                    let accounts = self.parse_action_argument(&method_name, &args);
-
-                    result.is_action_found = true;
-                    result.log.push(ActionResultLog {
-                        accounts: accounts.clone(),
-                        method: method_name,
-                    });
-                    result.accounts.extend(accounts);
-                }
-            }
-        }
-
-        result
-    }
-
-    /// Parse action arguments and return accounts
-    /// Main goal is gather all accounts that can be modified.
-    /// In `aurora-eth-connector` migration just receive balances
-    /// for modified accounts. So main goals is just catch modified
-    /// accounts.
-    ///
-    /// `deposit` function catch just for log information.
-    ///
-    /// `withdraw` function catch also for information, the `sender_id`
-    /// is from `predecessor_account_id` that we catch early in main flow.
-    ///
-    /// `storage_withdraw` - do nothing with account, just for logs.
-    ///
-    /// `storage_unregister` - use `predecessor_account_id` (that catch in
-    /// other flow), just for log.
-    #[must_use]
-    pub fn parse_action_argument(&self, method: &str, args: &[u8]) -> Vec<AccountId> {
-        use serde::Deserialize;
-
-        match method {
-            "ft_transfer" => {
-                #[derive(Debug, Deserialize)]
-                pub struct FtTransferArgs {
-                    pub receiver_id: AccountId,
-                    #[allow(dead_code)]
-                    pub amount: U128,
-                    #[allow(dead_code)]
-                    pub memo: Option<String>,
-                }
-                if let Ok(res) = serde_json::from_slice::<FtTransferArgs>(args) {
-                    print_log("ft_transfer");
-                    vec![res.receiver_id]
-                } else {
-                    print_log(" Failed deserialize FtTransferArgs");
-                    vec![]
-                }
-            }
-            "ft_transfer_call" => {
-                #[derive(Debug, Deserialize)]
-                pub struct FtTransferCallArgs {
-                    pub receiver_id: AccountId,
-                    #[allow(dead_code)]
-                    pub amount: U128,
-                    #[allow(dead_code)]
-                    pub memo: Option<String>,
-                    #[allow(dead_code)]
-                    pub msg: String,
-                }
-                if let Ok(res) = serde_json::from_slice::<FtTransferCallArgs>(args) {
-                    print_log("ft_transfer_call");
-                    vec![res.receiver_id]
-                } else {
-                    print_log("Failed deserialize FtTransferCallArgs");
-                    vec![]
-                }
-            }
-            "withdraw" => {
-                print_log(" Withdraw");
-                vec![]
-            }
-            "storage_deposit" => {
-                #[derive(Debug, Clone, Deserialize)]
-                pub struct StorageDepositArgs {
-                    pub account_id: Option<AccountId>,
-                    #[allow(dead_code)]
-                    pub registration_only: Option<bool>,
-                }
-                if let Ok(res) = serde_json::from_slice::<StorageDepositArgs>(args) {
-                    print_log("storage_deposit");
-                    if let Some(account_id) = res.account_id {
-                        vec![account_id]
-                    } else {
-                        vec![]
-                    }
-                } else {
-                    print_log("Failed deserialize FinishDepositArgs");
-                    vec![]
-                }
-            }
-            "storage_withdraw" => {
-                print_log("storage_withdraw");
-                vec![]
-            }
-            "storage_unregister" => {
-                print_log("storage_unregister");
-                vec![]
-            }
-            _ => vec![],
-        }
-    }
-
-    /// Get transactions and receipts indexed data from chunks.
-    /// Return indexed data including actions log.
-    ///
-    /// Special notice to catch `predecessor_account_id`:
-    /// It's especially important
-    /// for `withdraw`, `ft_transfer`, `ft_transfer_call`
-    /// and all `storage_deposit`,`storage_withdraw`,
-    /// `storage_unregister`.
-    pub async fn get_chunk_indexed_data(
-        &mut self,
-        chunks: Vec<ChunkHeaderView>,
-        block_height: BlockHeight,
-    ) -> IndexedData {
-        let mut results = IndexedData {
-            accounts: HashSet::new(),
-            logs: vec![],
-        };
-
-        // Fetch all chunks from block
-        for chunk in chunks {
-            // Get chunk data
-            let Ok(chunk_data) = self
-                .call(methods::chunk::RpcChunkRequest {
-                    chunk_reference:
-                        near_jsonrpc_primitives::types::chunks::ChunkReference::ChunkHash {
-                            chunk_id: chunk.chunk_hash,
-                        },
-                })
-                .await
-            else {
-                print_log("Failed get chunk");
-                // Set block as unresolved
-                self.unresolved_blocks.insert(block_height);
-                continue;
-            };
-
-            // Fetch chunk transactions
-            for tx in &chunk_data.transactions {
-                // We should process only specific receiver
-                if tx.receiver_id.as_str() != AURORA_CONTRACT {
-                    continue;
-                }
-                // Get actions from transaction
-                let res = self.get_actions_data(tx.actions.clone());
-
-                // Added predecessor account. It's especially important
-                // for `withdraw`, `ft_transfer`, `ft_transfer_call`
-                // and all `storage_deposit`,`storage_withdraw`,
-                // `storage_unregister`
-                if res.is_action_found {
-                    results
-                        .accounts
-                        .insert(AccountId::from_str(tx.signer_id.as_str()).unwrap());
-                    results.accounts.insert(AURORA_CONTRACT.parse().unwrap());
-
-                    let mut log = res.log;
-                    if !log.is_empty() {
-                        log[0]
-                            .accounts
-                            .push(AccountId::from_str(tx.signer_id.as_str()).unwrap());
-                        log[0].accounts.push(AURORA_CONTRACT.parse().unwrap());
-                    }
-                    results.logs.push(IndexedResultLog {
-                        block_height,
-                        actions: log,
-                    });
-                }
-                for account in res.accounts {
-                    results.accounts.insert(account);
-                }
-            }
-
-            // Fetch chunk transactions for receipts
-            for receipt in &chunk_data.receipts {
-                // We should process only specific receiver
-                if receipt.receiver_id.as_str() != AURORA_CONTRACT {
-                    continue;
-                }
-
-                // Get actions accounts from receipt
-                if let near_primitives::views::ReceiptEnumView::Action {
-                    signer_id, actions, ..
-                } = receipt.receipt.clone()
-                {
-                    let res = self.get_actions_data(actions);
-                    // Added predecessor_account_id.
-                    // NOTE: same notice as before about importance
-                    // for that field.
-                    if res.is_action_found {
-                        results
-                            .accounts
-                            .insert(AccountId::from_str(signer_id.as_str()).unwrap());
-                        results
-                            .accounts
-                            .insert(AccountId::from_str(receipt.predecessor_id.as_str()).unwrap());
-                        results
-                            .accounts
-                            .insert(AccountId::from_str(receipt.receiver_id.as_str()).unwrap());
-
-                        let mut log = res.log;
-                        if !log.is_empty() {
-                            log[0]
-                                .accounts
-                                .push(AccountId::from_str(signer_id.as_str()).unwrap());
-                            log[0].accounts.push(
-                                AccountId::from_str(receipt.predecessor_id.as_str()).unwrap(),
-                            );
-                            log[0]
-                                .accounts
-                                .push(AccountId::from_str(receipt.receiver_id.as_str()).unwrap());
-                        }
-                        results.logs.push(IndexedResultLog {
-                            block_height,
-                            actions: log,
-                        });
-                    }
-                    for account in res.accounts {
-                        results.accounts.insert(account);
-                    }
-                }
-            }
-        }
-        // Flow passed successfully - remove block
-        if self.unresolved_blocks.contains(&block_height) {
-            self.unresolved_blocks.remove(&block_height);
-        }
-        results
     }
 
     /// Commit transaction and wait respond. It should retry if it's fail
